trading_research_lab/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trading-research-lab"
version = "0.1.0"
description = "Sane Self-Improving Trading Research Lab"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.1",
  "pyarrow>=14.0",
  "PyYAML>=6.0",
  "requests>=2.31",
]

[project.optional-dependencies]
dev = ["pytest>=7.4"]

[project.scripts]
lab = "lab.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
